Using decimal for sizes instead of doubles

+++ SourceCode/Common/Extensions/LongExtensions.cs
@@ -8,19 +8,19 @@ namespace Common.Extensions
 {
     public static class LongExtensions
     {
-        public static double GetKB(this long size)
+        public static decimal GetKB(this long size)
         {
-            return ((double)size) / 1024;
+            return ((decimal)size) / 1024;
         }
 
-        public static double GetMB(this long size)
+        public static decimal GetMB(this long size)
         {
-            return ((double)size) / 1024 / 1024;
+            return ((decimal)size) / 1024 / 1024;
         }
 
-        public static double GetGB(this long size)
+        public static decimal GetGB(this long size)
         {
-            return ((double)size) / 1024 / 1024 / 1024;
+            return ((decimal)size) / 1024 / 1024 / 1024;
         }
     }
 }
